로그인 화면 UI 수정 : refactor : 로그인 관련 화면, 함수, LoginPlatform  리팩토링https://github.com/Capstone-4Potato/frontend/issues/1
