--- lib/colors.dart
@@ -1,8 +1,0 @@
-import 'package:flutter/material.dart';
-
-Color primary = const Color.fromARGB(255, 230, 105, 50);
-Color background = const Color.fromARGB(255, 245, 245, 245);
-Color bam = const Color.fromARGB(255, 73, 43, 33);
-Color accent = const Color.fromARGB(255, 242, 102, 71);
-Color progress_color = const Color.fromARGB(255, 255, 219, 181);
-Color back_progress_color = const Color.fromARGB(255, 237, 148, 109);
