+++ lib/learning_coures/sentecnes/sentencelearningcard.dart
@@ -3,7 +3,7 @@ import 'dart:convert';
 import 'dart:io';
 import 'package:audioplayers/audioplayers.dart';
 import 'package:flutter/material.dart';
-import 'package:flutter_application_1/colors.dart';
+import 'package:flutter_application_1/new/models/colors.dart';
 import 'package:flutter_application_1/learning_coures/sentecnes/sentencefeedbackui.dart';
 import 'package:flutter_application_1/home/home_nav.dart';
 import 'package:flutter_application_1/widgets/exit_dialog.dart';
