--- lib/login/login_platform.dart
@@ -1,7 +1,0 @@
-enum LoginPlatform {
-  google,
-  kakao,
-  naver,
-  apple,
-  none,
-}
