+++ lib/new/models/colors.dart
@@ -1,0 +1,21 @@
+import 'package:flutter/material.dart';
+
+Color primary = const Color.fromARGB(255, 230, 105, 50);
+Color background = const Color.fromARGB(255, 245, 245, 245);
+Color bam = const Color.fromARGB(255, 73, 43, 33);
+Color accent = const Color.fromARGB(255, 242, 102, 71);
+Color progress_color = const Color.fromARGB(255, 255, 219, 181);
+Color back_progress_color = const Color.fromARGB(255, 237, 148, 109);
+
+class AppColors {
+  static const Color primary = Color(0xFFE66932);
+  static const Color background = Color(0xFFFFFFFF);
+  static const Color iconColor = Color(0xFF492B21);
+
+  static const Color brown = Color(0xFF492B21);
+  static const Color black = Color(0xFF000000);
+  static const Color yellow = Color(0xFFFEE500);
+  static const Color white = Color(0xFFFFFFFF);
+  static const Color lightBrown = Color.fromRGBO(73, 43, 33, 0.7); // 밝은 갈색
+  static const Color lightOrange = Color(0xFFF26647); // 밝은 주황
+}
