+++ lib/new/models/font_family.dart
@@ -1,0 +1,9 @@
+/// 폰트 이름 저장
+enum FontFamily {
+  bmJua(fontName: 'BM_Jua'),
+  marine(fontName: 'Marine');
+
+  final String fontName;
+
+  const FontFamily({required this.fontName});
+}
