+++ lib/new/models/shared_preferences_key.dart
@@ -1,0 +1,3 @@
+enum SharedPreferencesKey {
+  loginPlatform,
+}
