+++ lib/settings/profile_page.dart
@@ -1,14 +1,14 @@
 import 'package:flutter/material.dart';
-import 'package:flutter_application_1/colors.dart';
+import 'package:flutter_application_1/new/models/colors.dart';
 import 'package:flutter_application_1/icons/custom_icons.dart';
-import 'package:flutter_application_1/login/login_platform.dart';
+import 'package:flutter_application_1/new/models/login_platform.dart';
 import 'package:flutter_application_1/main.dart';
 import 'package:flutter_application_1/settings/editprofile/editprofile_screen.dart';
 import 'package:flutter_application_1/tutorial/retutorial.dart';
 import 'package:flutter_application_1/settings/logout/sign_out_social.dart';
 import 'package:flutter_application_1/settings/logout/signout.dart';
 import 'package:flutter_application_1/settings/deleteaccount/withdrawal_screen.dart';
-import 'package:flutter_application_1/login/login_screen.dart';
+import 'package:flutter_application_1/new/screens/login_screen.dart';
 import 'package:flutter_application_1/userauthmanager.dart';
 import 'package:flutter_screenutil/flutter_screenutil.dart';
 import 'dart:convert';
