+++ lib/tutorial/learning_course_tutorial_screen2.dart
@@ -1,6 +1,6 @@
 import 'package:dotted_decoration/dotted_decoration.dart';
 import 'package:flutter/material.dart';
-import 'package:flutter_application_1/colors.dart';
+import 'package:flutter_application_1/new/models/colors.dart';
 import 'package:flutter_screenutil/flutter_screenutil.dart';
 import 'package:glowy_borders/glowy_borders.dart';
