+++ lib/widgets/audio_graph.dart
@@ -1,7 +1,7 @@
 import 'dart:math';
 import 'package:fl_chart/fl_chart.dart';
 import 'package:flutter/material.dart';
-import 'package:flutter_application_1/colors.dart';
+import 'package:flutter_application_1/new/models/colors.dart';
 import 'package:flutter_application_1/feedback_data.dart';
 import 'package:flutter_screenutil/flutter_screenutil.dart';
