+++ lib/widgets/exit_dialog.dart
@@ -1,5 +1,5 @@
 import 'package:flutter/material.dart';
-import 'package:flutter_application_1/colors.dart';
+import 'package:flutter_application_1/new/models/colors.dart';
 import 'package:flutter_application_1/home/home_nav.dart';
 import 'package:flutter_screenutil/flutter_screenutil.dart';
