+++ lib/widgets/recording_error_dialog.dart
@@ -1,5 +1,5 @@
 import 'package:flutter/material.dart';
-import 'package:flutter_application_1/colors.dart';
+import 'package:flutter_application_1/new/models/colors.dart';
 import 'package:flutter_screenutil/flutter_screenutil.dart';
 
 class RecordingErrorDialog extends StatelessWidget {
